invitations_views: Refactor tests
DRY
